requirements.txt
click==8.1.7
explain==0.1.1
keras==2.12.0
matplotlib==3.7.2
numpy==2.1.2
pandas==2.2.3
plotly==5.19.0
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.3.0
scipy==1.14.1
seaborn==0.13.2
setuptools==69.2.0
tensorflow==2.12.1
utils==1.0.2
